refactor : 예외 검증 방식 변경
