+++ src/main/java/menu/domain/InedibleMenus.java
@@ -5,7 +5,6 @@ import java.util.List;
 import menu.util.StringConvertor;
 
 public class InedibleMenus {
-    private static final int MAX_MENU_COUNT = 2;
     private final List<String> menus;
 
     private InedibleMenus(List<String> menus) {
@@ -13,23 +12,9 @@ public class InedibleMenus {
     }
 
     public static InedibleMenus from(String menuNames) {
-        validateMenuName(menuNames);
-        String[] names = StringConvertor.splitByComma(menuNames);
-        validateMenuCount(names);
-        return new InedibleMenus(Arrays.asList(names));
+        return new InedibleMenus(Arrays.asList(StringConvertor.splitByComma(menuNames)));
     }
 
-    private static void validateMenuName(String input) {
-        if (!Menu.containsMenu(input)) {
-            throw new IllegalArgumentException("메뉴판에 없는 메뉴 입니다.");
-        }
-    }
-
-    private static void validateMenuCount(String[] names) {
-        if (names.length > MAX_MENU_COUNT) {
-            throw new IllegalArgumentException("먹지 못하는 메뉴는 최대 2개 까지 입력 가능합니다.")
-        }
-    }
 
     @Override
     public String toString() {
