+++ src/main/java/menu/view/InputView.java
@@ -2,20 +2,21 @@ package menu.view;
 
 import camp.nextstep.edu.missionutils.Console;
 import menu.util.StringConvertor;
-import menu.view.validation.InputValidator;
+import menu.view.validation.CoachNameValidator;
+import menu.view.validation.InedibleMenuValidator;
 
 public class InputView {
 
     public String inputCoaches() {
         String input = Console.readLine();
         String trimedInput = StringConvertor.convertTrimedString(input);
-        InputValidator.validateNameSeparator(trimedInput);
+        CoachNameValidator.validate(trimedInput);
         return trimedInput;
     }
 
     public String inputInedibleMenus() {
         String input = Console.readLine();
-        InputValidator.validateMenuSeparator(input);
+        InedibleMenuValidator.validate(input);
         return input;
     }
 }
